pyproject.toml
[tool.poetry]
name = "kaiano"
version = "0.1.25"
description = "Common utility functions shared across Kaiano’s projects."
authors = ["Kaiano Levine <kaiano@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "kaiano", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
gspread = "^6.0.2"
google-auth = "^2.29.0"
google-api-python-client = "^2.177.0"
spotipy = "^2.23.0"
python-dotenv = "^1.0.1"
mutagen = "^1.47.0"
pydrive2 = "^1.16.0"
pydrive = "^1.3.1"
gspread-formatting = "^1.2.1"
pytz = "^2025.2"
music-tag = "^0.4.3"
pyacoustid = "^1.3.0"
musicbrainzngs = "^0.7.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest = "^8.4.0"
pytest-cov = "^6.2.1"
pre-commit = "^4.2.0"
coverage = "^7.10.1"
isort = "^5.13.2"
mypy = "^1.11.2"
pytz = "^2025.2"
types-pytz = "^2025.2.0.20250809"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["kaiano"]

[tool.coverage.report]
fail_under = 90
show_missing = true
